Move pmk code under pmk package rather than core package
+ Fix Makefile
