+++ pkg/pmk/types.go
@@ -1,4 +1,4 @@
-package core
+package pmk
 
 import (
 	"net"
