--- pkg/util/keystone.go
@@ -1,1 +1,0 @@
-package util
